Rename controller to human activity.
