+++ TelepresenceBot/mobile/src/main/java/com/novoda/tpbot/human/ControllerView.java
@@ -1,4 +1,4 @@
-package com.novoda.tpbot.controller;
+package com.novoda.tpbot.human;
 
 import android.content.Context;
 import android.util.AttributeSet;
@@ -25,10 +25,13 @@ public class ControllerView extends RelativeLayout {
 
         View up = Views.findById(this, R.id.controller_up_button);
         up.setOnTouchListener(onButtonTouchListener);
+
         View down = Views.findById(this, R.id.controller_down_button);
         down.setOnTouchListener(onButtonTouchListener);
+
         View left = Views.findById(this, R.id.controller_left_button);
         left.setOnTouchListener(onButtonTouchListener);
+
         View right = Views.findById(this, R.id.controller_right_button);
         right.setOnTouchListener(onButtonTouchListener);
