+++ TelepresenceBot/mobile/src/main/java/com/novoda/tpbot/human/Direction.java
@@ -1,4 +1,4 @@
-package com.novoda.tpbot.controller;
+package com.novoda.tpbot.human;
 
 enum Direction {
     FORWARD("↑"),
